#197 add sport, useremail and status to card
