+++ src/components/ChooseApplication.tsx
@@ -47,10 +47,12 @@ export const ChooseApplication = () => {
             });
           }
         }
-        setSubmittedApplicationIDs(submittedApplicationIDs);
-        setInProgressApplicationIDs(inProgressApplicationIDs);
+        // setSubmittedApplicationIDs(submittedApplicationIDs);
+        // setInProgressApplicationIDs(inProgressApplicationIDs);
       }
     }
+    setSubmittedApplicationIDs(submittedApplicationIDs);
+    setInProgressApplicationIDs(inProgressApplicationIDs);
   }
 
   const renderSubmittedApplications = () => {
@@ -60,7 +62,6 @@ export const ChooseApplication = () => {
         to="/application"
         applicationId={applicationID.id}
         collectionName={applicationID.collection}
-        user={currentUser!.uid}
       ></AppCard>
     ));
   };
@@ -72,7 +73,6 @@ export const ChooseApplication = () => {
         to="/edit"
         applicationId={applicationID.id}
         collectionName={applicationID.collection}
-        user={currentUser!.email || ""}
       ></AppCard>
     ));
   };
